Allow Leg to 8
git-svn-id: file:///home/svn/DVA2006@2995 1e6d2d28-e1fb-0310-b8ab-81908d8c6b6a

+++ src/test/org/deltava/beans/schedule/TestScheduleEntry.java
@@ -80,7 +80,7 @@ public class TestScheduleEntry extends AbstractBeanTestCase {
 
    public void testValidation() {
       validateInput("leg", new Integer(-1), IllegalArgumentException.class);
-      validateInput("leg", new Integer(7), IllegalArgumentException.class);
+      validateInput("leg", new Integer(9), IllegalArgumentException.class);
       validateInput("flightNumber", new Integer(-1), IllegalArgumentException.class);
       validateInput("ID", new Integer(1), UnsupportedOperationException.class);
       validateInput("length", new Integer(-1), IllegalArgumentException.class);
